(GH-877) Add new columns to database for package
This includes a new PakageScanFlagResult which will hold information
about how "bad" a flagged scanned result was, i.e. note, warning, or
error.  These will control when a package is pushed back to the
maintainer to address.  Also added new columns for containing a reason
for an exemption from package validator and package scanner.